+++ R/derive_locf_records.R
@@ -15,8 +15,10 @@
 #'
 #' @param by_vars Grouping variables
 #'
-#'   For each group defined by `by_vars` an observation is added to the output
-#'   dataset. Only variables specified in `by_vars` will be populated
+#'   For each group defined by `by_vars` those observations from `dataset_expected_obs`
+#'   are added to the output dataset which do not have a corresponding observation
+#'   in the input dataset or for which `AVAL` is NA for the corresponding observation
+#'   in the input dataset. Only variables specified in `by_vars` will be populated
 #'   in the newly created records.
 #'
 #' @param order List of variables for sorting a dataset
@@ -27,92 +29,88 @@
 #' @author G Gayatri
 #'
 #' @details For each group (with respect to the variables specified for the
-#'   `by_vars` parameter) an observation is added to the output dataset from
-#'   `dataset_expected_obs`, if the input dataset does not contain observations
-#'   for missed visits/time points or if AVAL is missing.
+#' by_vars parameter) those observations from dataset_expected_obs are added to
+#' the output dataset
+#' - which do not have a corresponding observation in the input dataset or
+#' - for which `AVAL` is NA for the corresponding observation in the input dataset.
 #'
 #'   For the new observations, `AVAL` is set to the non-missing `AVAL` of the
 #'   previous observation in the input dataset (when sorted by `order`) and
-#'   `DTYPE` is set to `LOCF`.
+#'   `DTYPE` is set to "LOCF".
 #'
-#' @return The input dataset with the new `LOCF` observations added for each
+#' @return The input dataset with the new "LOCF" observations added for each
 #' `by_vars`. Note, a variable will only be populated in the new parameter rows
 #' if it is specified in `by_vars`.
 #'
-#' @keywords der_bds_findings
-#' @family der_bds_findings
+#' @keywords der_prm_bds_findings
+#' @family der_prm_bds_findings
 #'
 #' @export
 #'
 #' @examples
 #'
 #' advs <- tibble::tribble(
-#'   ~STUDYID,  ~USUBJID,      ~PARAMCD, ~PARAM,                          ~AVAL, ~AVISITN, ~AVISIT,
-#'   "CDISC01", "01-701-1015", "DIABP", "Diastolic Blood Pressure (mmHg)", 51,    0,       "BASELINE",
-#'   "CDISC01", "01-701-1015", "DIABP", "Diastolic Blood Pressure (mmHg)", 50,    2,       "WEEK 2",
-#'   "CDISC01", "01-701-1015", "DIABP", "Diastolic Blood Pressure (mmHg)", 51,    4,       "WEEK 4",
-#'   "CDISC01", "01-701-1015", "DIABP", "Diastolic Blood Pressure (mmHg)", 50,    6,       "WEEK 6",
-#'   "CDISC01", "01-701-1015", "SYSBP", "Systolic Blood Pressure (mmHg)", 121,    0,       "BASELINE",
-#'   "CDISC01", "01-701-1015", "SYSBP", "Systolic Blood Pressure (mmHg)", 121,    2,       "WEEK 2",
-#'   "CDISC01", "01-701-1015", "SYSBP", "Systolic Blood Pressure (mmHg)", 121,    4,       "WEEK 4",
-#'   "CDISC01", "01-701-1015", "SYSBP", "Systolic Blood Pressure (mmHg)", 121,    6,       "WEEK 6",
-#'   "CDISC01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", 79,    0,       "BASELINE",
-#'   "CDISC01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", 80,    2,       "WEEK 2",
-#'   "CDISC01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", NA,    4,       "WEEK 4",
-#'   "CDISC01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", NA,    6,       "WEEK 6",
-#'   "CDISC01", "01-701-1028", "SYSBP", "Systolic Blood Pressure (mmHg)", 130,    0,       "BASELINE",
-#'   "CDISC01", "01-701-1028", "SYSBP", "Systolic Blood Pressure (mmHg)", 132,    2,       "WEEK 2"
+#'   ~STUDYID, ~USUBJID, ~PARAMCD, ~AVAL, ~AVISITN, ~AVISIT,
+#'   "CDISC01", "01-701-1015", "PULSE", 61, 0, "BASELINE",
+#'   "CDISC01", "01-701-1015", "PULSE", 60, 2, "WEEK 6",
+#'   "CDISC01", "01-701-1015", "DIABP", 51, 0, "BASELINE",
+#'   "CDISC01", "01-701-1015", "DIABP", 50, 2, "WEEK 2",
+#'   "CDISC01", "01-701-1015", "DIABP", 51, 4, "WEEK 4",
+#'   "CDISC01", "01-701-1015", "DIABP", 50, 6, "WEEK 6",
+#'   "CDISC01", "01-701-1015", "SYSBP", 121, 0, "BASELINE",
+#'   "CDISC01", "01-701-1015", "SYSBP", 121, 2, "WEEK 2",
+#'   "CDISC01", "01-701-1015", "SYSBP", 121, 4, "WEEK 4",
+#'   "CDISC01", "01-701-1015", "SYSBP", 121, 6, "WEEK 6",
+#'   "CDISC01", "01-701-1028", "PULSE", 65, 0, "BASELINE",
+#'   "CDISC01", "01-701-1028", "DIABP", 79, 0, "BASELINE",
+#'   "CDISC01", "01-701-1028", "DIABP", 80, 2, "WEEK 2",
+#'   "CDISC01", "01-701-1028", "DIABP", NA, 4, "WEEK 4",
+#'   "CDISC01", "01-701-1028", "DIABP", NA, 6, "WEEK 6",
+#'   "CDISC01", "01-701-1028", "SYSBP", 130, 0, "BASELINE",
+#'   "CDISC01", "01-701-1028", "SYSBP", 132, 2, "WEEK 2"
 #' )
 #'
 #'
 #' # A dataset with all the combinations of PARAMCD, PARAM, AVISIT, AVISITN, ... which are expected.
 #' advs_expected_obsv <- tibble::tribble(
-#'  ~PARAMCD, ~PARAM, ~AVISITN, ~AVISIT,
-#'   "DIABP", "Diastolic Blood Pressure (mmHg)", 0, "BASELINE",
-#'   "DIABP", "Diastolic Blood Pressure (mmHg)", 2, "WEEK 2",
-#'   "DIABP", "Diastolic Blood Pressure (mmHg)", 4, "WEEK 4",
-#'   "DIABP", "Diastolic Blood Pressure (mmHg)", 6, "WEEK 6",
-#'   "SYSBP", "Systolic Blood Pressure (mmHg)",  0, "BASELINE",
-#'   "SYSBP", "Systolic Blood Pressure (mmHg)",  2, "WEEK 2",
-#'   "SYSBP", "Systolic Blood Pressure (mmHg)",  4, "WEEK 4",
-#'   "SYSBP", "Systolic Blood Pressure (mmHg)",  6, "WEEK 6"
+#'   ~PARAMCD, ~AVISITN, ~AVISIT,
+#'   "PULSE", 0, "BASELINE",
+#'   "PULSE", 0, "WEEK 6",
+#'   "DIABP", 0, "BASELINE",
+#'   "DIABP", 2, "WEEK 2",
+#'   "DIABP", 4, "WEEK 4",
+#'   "DIABP", 6, "WEEK 6",
+#'   "SYSBP", 0, "BASELINE",
+#'   "SYSBP", 2, "WEEK 2",
+#'   "SYSBP", 4, "WEEK 4",
+#'   "SYSBP", 6, "WEEK 6"
 #' )
 #'
-#' # Add a new record for each 'by_vars' when the input dataset does not contain
-#' # observations for missed visits/timepoints or when AVAL is missing.
-#' # Set DTYPE = 'LOCF' for the newly added records.
-#' # Carry forward the AVAL of the previous observation after sorting the dataset
-#' # by 'order' parameter.
-#' Final_advs <- derive_locf_records(data=advs,
-#'                                   dataset_expected_obs=advs_expected_obsv,
-#'                                   by_vars=vars(STUDYID, USUBJID, PARAM, PARAMCD),
-#'                                   order=vars(STUDYID, USUBJID, PARAM, PARAMCD, AVISITN, AVISIT))
-
-
-
+#' derive_locf_records(
+#'   data = advs,
+#'   dataset_expected_obs = advs_expected_obsv,
+#'   by_vars = vars(STUDYID, USUBJID, PARAMCD),
+#'   order = vars(STUDYID, USUBJID, PARAMCD, AVISITN, AVISIT)
+#' )
+#'
 derive_locf_records <- function(dataset,
                                 dataset_expected_obs,
                                 by_vars,
                                 order) {
-  library(tidyr)
-
   #### Input Checking ####
-
-  # Check by_vars and order variables in input dataset
-  assert_data_frame(dataset, required_vars = by_vars)
-  assert_data_frame(dataset, required_vars = order)
 
   # Check if input parameters is a valid list of variables
   assert_vars(by_vars, optional = TRUE)
   assert_order_vars(order)
 
-
+  # Check by_vars and order variables in input datasets
+  assert_data_frame(dataset_expected_obs)
+  assert_data_frame(
+    dataset,
+    required_vars = quo_c(by_vars, extract_vars(order), chr2vars(colnames(dataset_expected_obs)))
+  )
 
   #### Prepping 'dataset_expected_obs' ####
-
-  # Add temp_dtype='LOCF' for the 'expected observation dataset'
-  dataset_expected_obs <- dataset_expected_obs %>%
-    mutate(temp_dtype="LOCF")
 
 
   # Get the STUDYID and USUBJID from input dataset to set expected
@@ -138,7 +136,6 @@ derive_locf_records <- function(dataset,
 
   # Get the variable names from the expected observation dataset
   exp_obs_vars <- exp_obsv %>%
-    select(-temp_dtype) %>%
     colnames()
 
 
@@ -146,39 +143,46 @@ derive_locf_records <- function(dataset,
   # from the original input dataset (with non-missing AVAL)
   advs_unique_original <- aval_not_missing %>%
     select(all_of(exp_obs_vars)) %>%
-    drop_na() %>%
     distinct()
 
+
+  tmp_dtype <- get_new_tmp_var(exp_obsv, prefix = "tmp_dtype")
 
   # Get all the expected observations that are to be added to the input
   # dataset (with non-missing AVAL)
   advs_exp_obsv3 <- exp_obsv %>%
-    anti_join(advs_unique_original, by=c(exp_obs_vars))
+    mutate(!!tmp_dtype := "LOCF") %>%
+    anti_join(advs_unique_original, by = c(exp_obs_vars)) %>%
+    drop_na(exp_obs_vars)
 
 
   # Merge the expected observations with the input dataset (with non-missing AVAL)
   # Arrange the dataset by 'order' and group it by 'by_vars'
   # Use fill() to fill the AVAL from the previous observation for the newly added records
-  aval_not_missing_locf <- aval_not_missing %>%
-    full_join(advs_exp_obsv3,  by=c(exp_obs_vars)) %>%
-    mutate(DTYPE = case_when(temp_dtype=="LOCF" ~ "LOCF")) %>%
-    select(-temp_dtype) %>%
-    arrange(!!!order) %>%
-    group_by(!!!by_vars) %>%
-    fill("AVAL") %>%
-    ungroup()
+
+  attach(aval_not_missing)
+
+  if (exists("DTYPE") == FALSE) {
+    aval_not_missing_locf <- aval_not_missing %>%
+      full_join(advs_exp_obsv3, by = c(exp_obs_vars)) %>%
+      mutate(DTYPE = case_when(!!tmp_dtype == "LOCF" ~ "LOCF")) %>%
+      select(-!!tmp_dtype) %>%
+      arrange(!!!order) %>%
+      group_by(!!!by_vars) %>%
+      fill("AVAL") %>%
+      ungroup()
+  } else if (exists("DTYPE") == TRUE) {
+    aval_not_missing_locf <- aval_not_missing %>%
+      full_join(advs_exp_obsv3, by = c(exp_obs_vars)) %>%
+      mutate(DTYPE = if_else(!!tmp_dtype == "LOCF", "LOCF", DTYPE, missing = DTYPE)) %>%
+      select(-!!tmp_dtype) %>%
+      arrange(!!!order) %>%
+      group_by(!!!by_vars) %>%
+      fill("AVAL") %>%
+      ungroup()
+  }
 
 
   # Output dataset - merge the AVAL missing with non-missing+newly added LOCF records
-   output <- bind_rows(aval_not_missing_locf, aval_missing) %>%
-    arrange(!!!order)
-
-
-
+  output <- bind_rows(aval_not_missing_locf, aval_missing)
 }
-
-
-
-
-
-