+++ NEWS.md
@@ -32,6 +32,10 @@ reference dataset. (#1477)
 
 - Removed R Version 3.6 check in CI/CD workflows in favor of the three most recent versions: 4.0, 4.1 and 4.2. (#1556)
 
+- The new function `derive_locf_records()` adds LOCF records as new observations. 
+This can be used when the input dataset does not contain observations for missed 
+visits/time points or when `AVAL` is NA for particular visits/time points (#1316).
+
 ## Updates of Existing Functions
 
 - Argument `ignore_time_for_ref_end_date` was added to `derive_var_ontrtfl()`,